# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cidade Aberta - Plataforma de Participação Cidadã
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service-orders:
    get:
      operationId: listServiceOrders
      tags: [service-orders]
      summary: List service orders (public/admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: teamId
          in: query
          schema:
            type: ["integer", "null"]
        - name: demandId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Service orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceOrder"

    post:
      operationId: createServiceOrder
      tags: [service-orders]
      summary: Create a service order from a demand (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
  /service-orders/{id}:
    get:
      operationId: getServiceOrder
      tags: [service-orders]
      summary: Get service order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"

    patch:
      operationId: updateServiceOrder
      tags: [service-orders]
      summary: Update service order status/team (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"

  # TEAMS
...
  /users/{id}/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get citizen profile and history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/ranking:
    get:
      operationId: getCitizenRanking
      tags: [users]
      summary: Get citizen participation ranking
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Ranking
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CitizenRankItem"
components:
  schemas:
...
    Demand:
      type: object
      required: [id, title, description, status, categoryId, neighborhoodId, address, lat, lng, confirmationsCount, commentsCount, createdAt, authorId]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, under_review, approved, in_progress, completed, rejected, closed]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, urgent, null]
        categoryId:
          type: integer
        category:
          $ref: "#/components/schemas/Category"
        neighborhoodId:
          type: integer
        neighborhood:
          $ref: "#/components/schemas/Neighborhood"
        departmentId:
          type: ["integer", "null"]
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        photoUrl:
          type: ["string", "null"]
        confirmationsCount:
          type: integer
        commentsCount:
          type: integer
        isConfirmedByMe:
          type: boolean
        authorId:
          type: integer
        authorName:
          type: string
        publicResponse:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Category:
      type: object
      required: [id, name, icon, color]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        description:
          type: ["string", "null"]
        demandsCount:
          type: integer
...
    Neighborhood:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        demandsCount:
          type: integer
        resolvedCount:
          type: integer

    # SERVICE ORDER
    ServiceOrder:
      type: object
      required: [id, demandId, status, createdAt]
      properties:
        id:
          type: integer
        demandId:
          type: integer
        demand:
          $ref: "#/components/schemas/Demand"
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        teamId:
          type: ["integer", "null"]
        team:
          $ref: "#/components/schemas/Team"
        estimatedCompletion:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        slaHours:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ServiceOrderInput:
      type: object
      required: [demandId, title, priority]
      properties:
        demandId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        teamId:
          type: ["integer", "null"]
        estimatedCompletion:
          type: ["string", "null"]
        slaHours:
          type: ["integer", "null"]
    ServiceOrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, cancelled]
        teamId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]

    # TEAM
    Team:
      type: object
      required: [id, name, area]
      properties:
        id:
          type: integer
        name:
          type: string
        area:
          type: string
        description:
          type: ["string", "null"]
        activeOrdersCount:
          type: integer
        completedOrdersCount:
          type: integer
...
    UserProfile:
      type: object
      required: [id, name, role, points, demandsCount, confirmationsCount, commentsCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        points:
          type: integer
        demandsCount:
          type: integer
        confirmationsCount:
          type: integer
        commentsCount:
          type: integer
        rankPosition:
          type: ["integer", "null"]
        recentDemands:
          type: array
          items:
            $ref: "#/components/schemas/Demand"
        createdAt:
          type: string
    CitizenRankItem:
      type: object
      required: [userId, name, points, demandsCount, confirmationsCount]
      properties:
        userId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        points:
          type: integer
        demandsCount:
          type: integer
        confirmationsCount:
          type: integer
        rank:
          type: integer